+++ influxql/select.go
@@ -287,7 +287,7 @@ func buildExprIterator(expr Expr, ic IteratorCreator, opt IteratorOptions) (Iter
 					if err != nil {
 						return nil, err
 					}
-					n := expr.Args[len(expr.Args)-1].(*NumberLiteral)
+					n := expr.Args[len(expr.Args)-1].(*IntegerLiteral)
 					return newTopIterator(input, opt, n, tags)
 				case "bottom":
 					var tags []int
@@ -311,14 +311,20 @@ func buildExprIterator(expr Expr, ic IteratorCreator, opt IteratorOptions) (Iter
 					if err != nil {
 						return nil, err
 					}
-					n := expr.Args[len(expr.Args)-1].(*NumberLiteral)
+					n := expr.Args[len(expr.Args)-1].(*IntegerLiteral)
 					return newBottomIterator(input, opt, n, tags)
 				case "percentile":
 					input, err := buildExprIterator(expr.Args[0].(*VarRef), ic, opt)
 					if err != nil {
 						return nil, err
 					}
-					percentile := expr.Args[1].(*NumberLiteral).Val
+					var percentile float64
+					switch arg := expr.Args[1].(type) {
+					case *NumberLiteral:
+						percentile = arg.Val
+					case *IntegerLiteral:
+						percentile = float64(arg.Val)
+					}
 					return newPercentileIterator(input, opt, percentile)
 				default:
 					return nil, fmt.Errorf("unsupported call: %s", expr.Name)
@@ -390,9 +396,14 @@ func buildRHSTransformIterator(lhs Iterator, rhs Literal, op Token, ic IteratorC
 			return nil, fmt.Errorf("type mismatch on LHS, unable to use %T as a FloatIterator", lhs)
 		}
 
-		lit, ok := rhs.(*NumberLiteral)
-		if !ok {
-			return nil, fmt.Errorf("type mismatch on RHS, unable to use %T as a NumberLiteral", lhs)
+		var val float64
+		switch rhs := rhs.(type) {
+		case *NumberLiteral:
+			val = rhs.Val
+		case *IntegerLiteral:
+			val = float64(rhs.Val)
+		default:
+			return nil, fmt.Errorf("type mismatch on RHS, unable to use %T as a NumberLiteral", rhs)
 		}
 		return &floatTransformIterator{
 			input: input,
@@ -400,8 +411,36 @@ func buildRHSTransformIterator(lhs Iterator, rhs Literal, op Token, ic IteratorC
 				if p == nil {
 					return nil
 				}
-				p.Value = fn(p.Value, lit.Val)
+				p.Value = fn(p.Value, val)
 				return p
+			},
+		}, nil
+	case func(int64, int64) float64:
+		input, ok := lhs.(IntegerIterator)
+		if !ok {
+			return nil, fmt.Errorf("type mismatch on LHS, unable to use %T as a IntegerIterator", lhs)
+		}
+
+		var val int64
+		switch rhs := rhs.(type) {
+		case *IntegerLiteral:
+			val = rhs.Val
+		default:
+			return nil, fmt.Errorf("type mismatch on RHS, unable to use %T as a IntegerLiteral", rhs)
+		}
+		return &integerFloatTransformIterator{
+			input: input,
+			fn: func(p *IntegerPoint) *FloatPoint {
+				if p == nil {
+					return nil
+				}
+				return &FloatPoint{
+					Name:  p.Name,
+					Tags:  p.Tags,
+					Time:  p.Time,
+					Value: fn(p.Value, val),
+					Aux:   p.Aux,
+				}
 			},
 		}, nil
 	case func(float64, float64) bool:
@@ -415,9 +454,14 @@ func buildRHSTransformIterator(lhs Iterator, rhs Literal, op Token, ic IteratorC
 			return nil, fmt.Errorf("type mismatch on LHS, unable to use %T as a FloatIterator", lhs)
 		}
 
-		lit, ok := rhs.(*NumberLiteral)
-		if !ok {
-			return nil, fmt.Errorf("type mismatch on RHS, unable to use %T as a NumberLiteral", lhs)
+		var val float64
+		switch rhs := rhs.(type) {
+		case *NumberLiteral:
+			val = rhs.Val
+		case *IntegerLiteral:
+			val = float64(rhs.Val)
+		default:
+			return nil, fmt.Errorf("type mismatch on RHS, unable to use %T as a NumberLiteral", rhs)
 		}
 		return &floatBoolTransformIterator{
 			input: input,
@@ -429,7 +473,64 @@ func buildRHSTransformIterator(lhs Iterator, rhs Literal, op Token, ic IteratorC
 					Name:  p.Name,
 					Tags:  p.Tags,
 					Time:  p.Time,
-					Value: fn(p.Value, lit.Val),
+					Value: fn(p.Value, val),
+					Aux:   p.Aux,
+				}
+			},
+		}, nil
+	case func(int64, int64) int64:
+		var input IntegerIterator
+		switch lhs := lhs.(type) {
+		case IntegerIterator:
+			input = lhs
+		default:
+			return nil, fmt.Errorf("type mismatch on LHS, unable to use %T as an IntegerIterator", lhs)
+		}
+
+		var val int64
+		switch rhs := rhs.(type) {
+		case *IntegerLiteral:
+			val = rhs.Val
+		default:
+			return nil, fmt.Errorf("type mismatch on RHS, unable to use %T as an IntegerLiteral", rhs)
+		}
+		return &integerTransformIterator{
+			input: input,
+			fn: func(p *IntegerPoint) *IntegerPoint {
+				if p == nil {
+					return nil
+				}
+				p.Value = fn(p.Value, val)
+				return p
+			},
+		}, nil
+	case func(int64, int64) bool:
+		var input IntegerIterator
+		switch lhs := lhs.(type) {
+		case IntegerIterator:
+			input = lhs
+		default:
+			return nil, fmt.Errorf("type mismatch on LHS, unable to use %T as an IntegerIterator", lhs)
+		}
+
+		var val int64
+		switch rhs := rhs.(type) {
+		case *IntegerLiteral:
+			val = rhs.Val
+		default:
+			return nil, fmt.Errorf("type mismatch on RHS, unable to use %T as an IntegerLiteral", rhs)
+		}
+		return &integerBoolTransformIterator{
+			input: input,
+			fn: func(p *IntegerPoint) *BooleanPoint {
+				if p == nil {
+					return nil
+				}
+				return &BooleanPoint{
+					Name:  p.Name,
+					Tags:  p.Tags,
+					Time:  p.Time,
+					Value: fn(p.Value, val),
 					Aux:   p.Aux,
 				}
 			},
@@ -452,8 +553,13 @@ func buildLHSTransformIterator(lhs Literal, rhs Iterator, op Token, ic IteratorC
 			return nil, fmt.Errorf("type mismatch on RHS, unable to use %T as a FloatIterator", rhs)
 		}
 
-		lit, ok := lhs.(*NumberLiteral)
-		if !ok {
+		var val float64
+		switch lhs := lhs.(type) {
+		case *NumberLiteral:
+			val = lhs.Val
+		case *IntegerLiteral:
+			val = float64(lhs.Val)
+		default:
 			return nil, fmt.Errorf("type mismatch on LHS, unable to use %T as a NumberLiteral", lhs)
 		}
 		return &floatTransformIterator{
@@ -462,8 +568,36 @@ func buildLHSTransformIterator(lhs Literal, rhs Iterator, op Token, ic IteratorC
 				if p == nil {
 					return nil
 				}
-				p.Value = fn(lit.Val, p.Value)
+				p.Value = fn(val, p.Value)
 				return p
+			},
+		}, nil
+	case func(int64, int64) float64:
+		input, ok := rhs.(IntegerIterator)
+		if !ok {
+			return nil, fmt.Errorf("type mismatch on RHS, unable to use %T as a IntegerIterator", lhs)
+		}
+
+		var val int64
+		switch lhs := lhs.(type) {
+		case *IntegerLiteral:
+			val = lhs.Val
+		default:
+			return nil, fmt.Errorf("type mismatch on LHS, unable to use %T as a IntegerLiteral", rhs)
+		}
+		return &integerFloatTransformIterator{
+			input: input,
+			fn: func(p *IntegerPoint) *FloatPoint {
+				if p == nil {
+					return nil
+				}
+				return &FloatPoint{
+					Name:  p.Name,
+					Tags:  p.Tags,
+					Time:  p.Time,
+					Value: fn(val, p.Value),
+					Aux:   p.Aux,
+				}
 			},
 		}, nil
 	case func(float64, float64) bool:
@@ -477,8 +611,13 @@ func buildLHSTransformIterator(lhs Literal, rhs Iterator, op Token, ic IteratorC
 			return nil, fmt.Errorf("type mismatch on RHS, unable to use %T as a FloatIterator", rhs)
 		}
 
-		lit, ok := lhs.(*NumberLiteral)
-		if !ok {
+		var val float64
+		switch lhs := lhs.(type) {
+		case *NumberLiteral:
+			val = lhs.Val
+		case *IntegerLiteral:
+			val = float64(lhs.Val)
+		default:
 			return nil, fmt.Errorf("type mismatch on LHS, unable to use %T as a NumberLiteral", lhs)
 		}
 		return &floatBoolTransformIterator{
@@ -491,7 +630,64 @@ func buildLHSTransformIterator(lhs Literal, rhs Iterator, op Token, ic IteratorC
 					Name:  p.Name,
 					Tags:  p.Tags,
 					Time:  p.Time,
-					Value: fn(lit.Val, p.Value),
+					Value: fn(val, p.Value),
+					Aux:   p.Aux,
+				}
+			},
+		}, nil
+	case func(int64, int64) int64:
+		var input IntegerIterator
+		switch rhs := rhs.(type) {
+		case IntegerIterator:
+			input = rhs
+		default:
+			return nil, fmt.Errorf("type mismatch on RHS, unable to use %T as an IntegerIterator", rhs)
+		}
+
+		var val int64
+		switch lhs := lhs.(type) {
+		case *IntegerLiteral:
+			val = lhs.Val
+		default:
+			return nil, fmt.Errorf("type mismatch on LHS, unable to use %T as an IntegerLiteral", lhs)
+		}
+		return &integerTransformIterator{
+			input: input,
+			fn: func(p *IntegerPoint) *IntegerPoint {
+				if p == nil {
+					return nil
+				}
+				p.Value = fn(val, p.Value)
+				return p
+			},
+		}, nil
+	case func(int64, int64) bool:
+		var input IntegerIterator
+		switch rhs := rhs.(type) {
+		case IntegerIterator:
+			input = rhs
+		default:
+			return nil, fmt.Errorf("type mismatch on RHS, unable to use %T as an IntegerIterator", rhs)
+		}
+
+		var val int64
+		switch lhs := lhs.(type) {
+		case *IntegerLiteral:
+			val = lhs.Val
+		default:
+			return nil, fmt.Errorf("type mismatch on LHS, unable to use %T as an IntegerLiteral", lhs)
+		}
+		return &integerBoolTransformIterator{
+			input: input,
+			fn: func(p *IntegerPoint) *BooleanPoint {
+				if p == nil {
+					return nil
+				}
+				return &BooleanPoint{
+					Name:  p.Name,
+					Tags:  p.Tags,
+					Time:  p.Time,
+					Value: fn(val, p.Value),
 					Aux:   p.Aux,
 				}
 			},
@@ -678,6 +874,8 @@ func literalDataType(lit Literal) DataType {
 	switch lit.(type) {
 	case *NumberLiteral:
 		return Float
+	case *IntegerLiteral:
+		return Integer
 	case *StringLiteral:
 		return String
 	case *BooleanLiteral: